refactor(bridge): move group info FFI export
